# garage/backend/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Мой гараж",
        "description": "Это список методов для интеграции с Мой гараж, в заголовках всех запросов должен присутствоваться API-ключ, который будет предоставлен администратором на почту клиента после регистрации клиента в сервисе администратором",
        "version": "1.0.2"
    },
    "paths": {
...
        "api/driver/upload-file": {
            "post": {
                "tags": [
                    "Files"
                ],
                "summary": "Загрузить файл",
                "description": "Загрузить файл и связать его с документами водителя аутентифицированного пользователя",
                "operationId": "uploadFile",
                "requestBody": {
                    "description": "Файл для загрузки",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "file": {
                                        "description": "Файл для загрузки",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "driverDocumentType": {
                                        "$ref": "#/components/schemas/DriverDocumentType"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Файл успешно загружен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "url": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Недопустимый тип файла"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DriverDocumentType": {
                "description": "The unique identifier of a product in our catalog",
                "type": null,
                "enum": [
                    "image_licence_front",
                    "image_licence_back",
                    "image_pasport_front",
                    "image_pasport_address",
                    "image_fase_and_pasport"
                ]
            },
...
        }
    }
}